backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
ultralytics==8.1.22
scikit-learn==1.4.1.post1
pymongo==4.6.2
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==23.2.1
websockets==12.0
httpx==0.27.0
pydantic==2.6.3
pydantic-settings==2.2.1
loguru==0.7.2
pytest==8.1.1
pytest-asyncio==0.23.5
